Luciferase-x5i.8: extract Culler (RDR-008 P4 — frustum/plane/ray bundled)
Phase 4 of the AbstractSpatialIndex god-class decomposition. Bundles the
frustum, plane, and ray clusters into a single Culler feature object per
RDR-008 §Decision item 2 (shared traversal hooks, high internal cohesion).

Pattern follows the P3 KnnSearcher extraction:
- New package lucien.cull with three files
- Culler<Key,ID,Content> owns 13 public cull entries + 12 helpers (frustum/
  plane/ray AABB/point/sphere intersection, entity-intersection calculators,
  AABB-closest-point helpers, ray-entity distance default, the standard
  non-DSOC cull fallback, and the getComponent slab accessor)
- CullGeometry<Key,ID,Content> consumer seam (10 methods: 7 subclass hooks
  + 2 cached entity accessors + spanning flag)
- FrustumCullProvider<Key,ID,Content> single-method producer seam Culler
  implements directly; DsocController consumes it as the standard-cull
  fallback when its Z-buffer is inactive

P2 substantive-critic obligation settled: FrustumGeometry SHRINKS by one
method (drops frustumCullVisibleStandard, re-homed in Culler).

DSOC ↔ standard-cull seam preserved: frustumCullVisible(frustum, cameraPos)
still branches dsoc != null ? dsoc.frustumCullVisible(...) :
culler.frustumCullVisibleStandard(...). The three composed derivatives
(frustumCullInside, frustumCullIntersecting, frustumCullWithinDistance)
compose on top of frustumCullVisible — never Culler directly — so they
inherit the DSOC routing decision. (Code-review-expert caught the initial
draft delegating these directly to Culler, silently bypassing DSOC; fix
applied with FrustumCullDsocRoutingTest as the regression net.)

DsocController ctor now takes FrustumCullProvider in addition to
FrustumGeometry; getStatistics() exposes dsocFrameCount/standardFrameCount
so the routing test has an observable hook.

this-escape invariant in the ctor extended to name CullGeometryImpl: any
feature object constructed in AbstractSpatialIndex's ctor must NOT invoke
methods on its supplied callback or on `this` during construction.

Subclasses (Octree, Tetree, Prism, SFCArrayIndex) byte-for-byte unchanged
— this is the LAST P-unchanged gate; P5 carries a bounded Tetree refactor.

Façade arc: AbstractSpatialIndex 4920 → 4109 lines (-811 net for P4, 5851
→ 4109 cumulative across P0+P1+P2+P3+P4, -1742 / -30%).

Tests: full lucien suite 2446/0/0/22 (was 2441, +5 from the new
FrustumCullDsocRoutingTest). Cull-focused sweep 169/0/0/7.

Reviews: code-review-expert + substantive-critic stacked per the standing
review-stacking feedback. code-review-expert: 0 Critical / 0 High after
the DSOC-bypass fix. substantive-critic: 0 Critical, 2 Significant (both
documentation — javadoc on the ID-only frustumCullVisible explaining the
intentional DSOC bypass, and dual inner-class duplication marker comments
on FrustumGeometryImpl + CullGeometryImpl). Both addressed in this PR.

Bead: Luciferase-x5i.8 (parent epic Luciferase-x5i). Next: G4 phase-
review-gate (Luciferase-x5i.9), then P5 collision (Luciferase-x5i.10).
